fix: rename folder for readability
